Refactored Unit & Integration tests
