+++ tests/integration/api/test_backend.py
@@ -9,14 +9,13 @@ from tests.defaults import CELERY_BACKEND_CLUSTER
 
 @pytest.mark.parametrize("node", [lazy_fixture(CELERY_BACKEND)])
 class test_celey_test_backend:
-    def test_ready(self, node: CeleryTestBackend):
-        assert node.ready()
-
     def test_app(self, node: CeleryTestBackend):
         assert node.app is None
 
 
 @pytest.mark.parametrize("cluster", [lazy_fixture(CELERY_BACKEND_CLUSTER)])
 class test_celery_backend_cluster:
-    def test_ready(self, cluster: CeleryBackendCluster):
-        assert cluster.ready()
+    def test_app(self, cluster: CeleryBackendCluster):
+        node: CeleryTestBackend
+        for node in cluster:
+            assert node.app is None
