+++ tests/integration/api/test_worker.py
@@ -12,9 +12,6 @@ from tests.integration.conftest import IntegrationWorkerContainer
 
 @pytest.mark.parametrize("node", [lazy_fixture(CELERY_WORKER)])
 class test_celey_test_worker:
-    def test_ready(self, node: CeleryTestWorker):
-        assert node.ready()
-
     def test_app(self, node: CeleryTestWorker, celery_setup_app: Celery):
         assert node.app is celery_setup_app
 
@@ -32,9 +29,6 @@ class test_celey_test_worker:
 
 @pytest.mark.parametrize("cluster", [lazy_fixture(CELERY_WORKER_CLUSTER)])
 class test_celery_worker_cluster:
-    def test_ready(self, cluster: CeleryWorkerCluster):
-        assert cluster.ready()
-
     def test_app(self, cluster: CeleryWorkerCluster, celery_setup_app: Celery):
         node: CeleryTestWorker
         for node in cluster:
