+++ tests/integration/vendors/test_redis.py
@@ -1,11 +1,12 @@
 import pytest
 from pytest_lazyfixture import lazy_fixture
 
-from pytest_celery import CELERY_REDIS_BACKEND
-from pytest_celery import CELERY_REDIS_BROKER
+# from pytest_celery import CELERY_REDIS_BACKEND
+# from pytest_celery import CELERY_REDIS_BROKER
 from pytest_celery import RedisContainer
-from pytest_celery import RedisTestBackend
-from pytest_celery import RedisTestBroker
+
+# from pytest_celery import RedisTestBackend
+# from pytest_celery import RedisTestBroker
 from tests.defaults import ALL_REDIS_FIXTURES
 
 
@@ -19,13 +20,13 @@ class test_redis_container:
         assert container.client.delete("ready")
 
 
-@pytest.mark.parametrize("node", [lazy_fixture(CELERY_REDIS_BACKEND)])
-class test_redis_test_backend:
-    def test_ready(self, node: RedisTestBackend):
-        assert node.ready()
+# @pytest.mark.parametrize("node", [lazy_fixture(CELERY_REDIS_BACKEND)])
+# class test_redis_test_backend:
+#     def test_placeholder(self, node: RedisTestBackend):
+#         node = node
 
 
-@pytest.mark.parametrize("node", [lazy_fixture(CELERY_REDIS_BROKER)])
-class test_redis_test_broker:
-    def test_ready(self, node: RedisTestBroker):
-        assert node.ready()
+# @pytest.mark.parametrize("node", [lazy_fixture(CELERY_REDIS_BROKER)])
+# class test_redis_test_broker:
+#     def test_placeholder(self, node: RedisTestBroker):
+#         node = node
