+++ tests/unit/api/test_worker.py
@@ -6,49 +6,25 @@ from pytest_celery import CeleryWorkerCluster
 
 
 class test_celey_test_worker:
-    def test_ready(self, unit_tests_container: CeleryTestContainer, celery_setup_app: Celery):
+    def test_app(self, unit_tests_container: CeleryTestContainer, celery_setup_app: Celery):
+        # TODO: Use mock instead of real container
         node = CeleryTestWorker(unit_tests_container, celery_setup_app)
-        assert node.ready()
-
-    def test_default_config_format(
-        self,
-        unit_tests_container: CeleryTestContainer,
-        celery_setup_app: Celery,
-    ):
-        node = CeleryTestWorker(unit_tests_container, celery_setup_app)
-        assert node.default_config() == dict()
+        assert node.app is celery_setup_app
 
     def test_version(self, unit_tests_container: CeleryTestContainer, celery_setup_app: Celery):
+        # TODO: Use mock instead of real container
         node = CeleryTestWorker(unit_tests_container, celery_setup_app)
         assert node.version == "unknown"
 
-    def test_wait_for_log(self, unit_tests_container: CeleryTestContainer, celery_setup_app: Celery):
-        node = CeleryTestWorker(unit_tests_container, celery_setup_app)
-        node.wait_for_log("", "test_celey_test_worker.test_wait_for_log")
-
-    def test_assert_log_exists(self, unit_tests_container: CeleryTestContainer, celery_setup_app: Celery):
-        node = CeleryTestWorker(unit_tests_container, celery_setup_app)
-        node.assert_log_exists("", "test_celey_test_worker.test_assert_log_exists")
-
 
 class test_celery_worker_cluster:
-    def test_ready(
-        self,
-        unit_tests_container: CeleryTestContainer,
-        local_test_container: CeleryTestContainer,
-        celery_setup_app: Celery,
-    ):
-        node1 = CeleryTestWorker(unit_tests_container, celery_setup_app)
-        node2 = CeleryTestWorker(local_test_container, celery_setup_app)
-        cluster = CeleryWorkerCluster(node1, node2)
-        assert cluster.ready()
-
     def test_app(
         self,
         unit_tests_container: CeleryTestContainer,
         local_test_container: CeleryTestContainer,
         celery_setup_app: Celery,
     ):
+        # TODO: Use mock instead of real container
         node1 = CeleryTestWorker(unit_tests_container, celery_setup_app)
         node2 = CeleryTestWorker(local_test_container, celery_setup_app)
         cluster = CeleryWorkerCluster(node1, node2)
@@ -56,23 +32,13 @@ class test_celery_worker_cluster:
         for node in cluster:
             assert node.app is celery_setup_app
 
-    def test_default_config_format(
-        self,
-        unit_tests_container: CeleryTestContainer,
-        local_test_container: CeleryTestContainer,
-        celery_setup_app: Celery,
-    ):
-        node1 = CeleryTestWorker(unit_tests_container, celery_setup_app)
-        node2 = CeleryTestWorker(local_test_container, celery_setup_app)
-        cluster = CeleryWorkerCluster(node1, node2)
-        assert cluster.default_config() == dict()
-
     def test_versions(
         self,
         unit_tests_container: CeleryTestContainer,
         local_test_container: CeleryTestContainer,
         celery_setup_app: Celery,
     ):
+        # TODO: Use mock instead of real container
         node1 = CeleryTestWorker(unit_tests_container, celery_setup_app)
         node2 = CeleryTestWorker(local_test_container, celery_setup_app)
         cluster = CeleryWorkerCluster(node1, node2)
