+++ tests/unit/conftest.py
@@ -1,4 +1,3 @@
-# from typing import Any
 from typing import Type
 
 import pytest
